requirement.txt
robotframework
robotframework-seleniumlibrary
webdrivermanager
robotframework-browser
pytest
PyYAML
